Create Inventory System
This code was written to be used in conjunction with a rasberryPi, an electronic striker and an RFID scanner.

The purpose of this project and proof of concept was to be used on a tool crib and would allow employees to scan their unique RFID cards to enter the toolcrib, while maintaining an inventory of what tools have been taken/returned. The system will further be used as an interior security system, integrating an SQL server ran off the companies private network.